gridmapapi/requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-extensions==2.2.8
idna==2.9
pip==9.0.3
psycopg2-binary==2.8.4
pycryptodome==3.9.7
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
setuptools==39.2.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
mozilla-django-oidc==1.2.3
django-logging-json
